Refactored occurrences utils by moving them all to new services module
It's not an util when it needs dependencies, so refactored occurrences utils by moving all the functions there to a new occurrences services module.

PT-1364.
